chore: moved repo finder integration test to new file
